requirements.txt
pydantic==1.8.2
envyaml==0.2060